project/backend/drweb_api/requirement.txt
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.2
image==1.5.27
Pillow==5.4.1
pytz==2018.9
